+++ source/App/source/ExampleHost.FunctionApp.Tests/Integration/ExampleHostTests.cs
@@ -38,6 +38,9 @@ namespace ExampleHost.FunctionApp.Tests.Integration
         {
             Fixture = fixture;
             Fixture.SetTestOutputHelper(testOutputHelper);
+
+            Fixture.App01HostManager.ClearHostLog();
+            Fixture.App02HostManager.ClearHostLog();
         }
 
         private ExampleHostFixture Fixture { get; }
@@ -52,6 +55,25 @@ namespace ExampleHost.FunctionApp.Tests.Integration
             Fixture.SetTestOutputHelper(null!);
 
             return Task.CompletedTask;
+        }
+
+        public static IEnumerable<object[]> TraceParentParameters()
+        {
+            yield return new object[]
+            {
+                TraceParentTestData.Empty,
+            };
+
+            yield return new object[]
+            {
+                new TraceParentTestData
+                {
+                     Version = "00",
+                     TraceId = Guid.NewGuid().ToString("N"),
+                     ParentId = string.Format("{0:x16}", new Random().Next(0x1000000)),
+                     TraceFlags = "01",
+                },
+            };
         }
 
         /// <summary>
@@ -98,20 +120,17 @@ namespace ExampleHost.FunctionApp.Tests.Integration
         ///
         /// 1: Both hosts must call "ConfigureFunctionsWorkerDefaults" with the following:
         /// <code>
-        ///     builder.UseMiddleware{CorrelationIdMiddleware}();
         ///     builder.UseMiddleware{FunctionTelemetryScopeMiddleware}();
         /// </code>
         ///
-        /// 2: Both hosts must call "ConfigureServices" with the following:
+        /// 2: Both hosts must call "ConfigureServices" with the following DataHub developed extension:
         /// <code>
-        ///     services.AddApplicationInsightsTelemetryWorkerService();
-        ///     services.AddScoped{ICorrelationContext, CorrelationContext}();
-        ///     services.AddScoped{CorrelationIdMiddleware}();
-        ///     services.AddScoped{FunctionTelemetryScopeMiddleware}();
+        ///     services.AddApplicationInsights();
         /// </code>
         /// </summary>
-        [Fact]
-        public async Task Middleware_Should_CauseExpectedEventsToBeLogged()
+        [Theory]
+        [MemberData(nameof(TraceParentParameters))]
+        public async Task Middleware_Should_CauseExpectedEventsToBeLogged(TraceParentTestData traceParentTestData)
         {
             var expectedEvents = new List<QueryResult>
             {
@@ -132,6 +151,11 @@ namespace ExampleHost.FunctionApp.Tests.Integration
             };
 
             using var request = new HttpRequestMessage(HttpMethod.Post, "api/v1/pet");
+            if (traceParentTestData != TraceParentTestData.Empty)
+            {
+                var traceParent = $"{traceParentTestData.Version}-{traceParentTestData.TraceId}-{traceParentTestData.ParentId}-{traceParentTestData.TraceFlags}";
+                request.Headers.Add("traceparent", traceParent);
+            }
 
             await Fixture.App01HostManager.HttpClient.SendAsync(request);
 
@@ -176,7 +200,7 @@ namespace ExampleHost.FunctionApp.Tests.Integration
                             queryTimerange);
 
                         actualCount = actualResponse.Value.Count;
-                        return ContainsExpectedEvents(expectedEvents, actualResponse.Value);
+                        return ContainsExpectedEvents(expectedEvents, actualResponse.Value, traceParentTestData);
                     },
                     waitLimit,
                     delay);
@@ -184,7 +208,7 @@ namespace ExampleHost.FunctionApp.Tests.Integration
             wasEventsLogged.Should().BeTrue($"'Was expected to log {expectedEvents.Count} number of events, but found {actualCount}. See log output for details.'");
         }
 
-        private bool ContainsExpectedEvents(IList<QueryResult> expectedEvents, IReadOnlyList<QueryResult> actualResults)
+        private bool ContainsExpectedEvents(IList<QueryResult> expectedEvents, IReadOnlyList<QueryResult> actualResults, TraceParentTestData traceParentTestData)
         {
             if (actualResults.Count < expectedEvents.Count)
             {
@@ -236,7 +260,15 @@ namespace ExampleHost.FunctionApp.Tests.Integration
                 }
             }
 
-            return true;
+            if (traceParentTestData == TraceParentTestData.Empty)
+            {
+                return true;
+            }
+
+            // If we added ´traceparent´ header while requesting, the Trace Id and Parent Id will be set for the first activity.
+            return actualResults.Any(actual =>
+                actual.OperationId == traceParentTestData.TraceId
+                && actual.ParentId == traceParentTestData.ParentId);
         }
 
         private static async Task AssertFunctionExecuted(FunctionAppHostManager hostManager, string functionName)
@@ -262,6 +294,24 @@ namespace ExampleHost.FunctionApp.Tests.Integration
         private void AssertNoExceptionsThrown()
         {
             Fixture.App01HostManager.CheckIfFunctionThrewException().Should().BeFalse();
+        }
+
+        public record TraceParentTestData
+        {
+            public static TraceParentTestData Empty { get; }
+                = new TraceParentTestData();
+
+            public string Version { get; set; }
+                = string.Empty;
+
+            public string TraceId { get; set; }
+                = string.Empty;
+
+            public string ParentId { get; set; }
+                = string.Empty;
+
+            public string TraceFlags { get; set; }
+                = string.Empty;
         }
 
         private class QueryResult