+++ source/App/source/ExampleHost.FunctionApp01/Program.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Azure.Messaging.ServiceBus;
+using Energinet.DataHub.Core.App.FunctionApp.Extensions.DependencyInjection;
 using Energinet.DataHub.Core.App.FunctionApp.FunctionTelemetryScope;
 using ExampleHost.FunctionApp01.Common;
 using Microsoft.Extensions.DependencyInjection;
@@ -33,13 +34,10 @@ var host = new HostBuilder()
         //  - https://github.com/Azure/azure-functions-dotnet-worker/issues/760
         //  - https://github.com/Azure/azure-functions-dotnet-worker/issues/822#issuecomment-1088012705
 
-        //// UNDONE: Track custom operations with App Insights SDK's [https://docs.microsoft.com/en-us/azure/azure-monitor/app/custom-operations-tracking]
-
         // CONCLUSION: We can use ILogger<> without calling the following:
         ////services.AddLogging();
 
-        services.AddApplicationInsightsTelemetryWorkerService();
-        services.AddScoped<FunctionTelemetryScopeMiddleware>();
+        services.AddApplicationInsights();
 
         services.AddSingleton(_ =>
         {